feat: add response to deck and fixed patch deck

+++ backend/internal/services/decks.go
@@ -6,7 +6,9 @@ import (
 	"memora/internal/errors"
 	"memora/internal/firebase"
 	"memora/internal/models"
+	"memora/internal/utils"
 
+	"cloud.google.com/go/firestore"
 	"github.com/go-playground/validator/v10"
 )
 
@@ -20,8 +22,8 @@ func NewDeckService(repo firebase.DeckRepository, validate *validator.Validate) 
 }
 
 func (s *DeckService) RegisterNewDeck(ctx context.Context, deck models.CreateDeck) (string, error) {
-	if deck.SharedEmails == nil {
-		deck.SharedEmails = []string{}
+	if err := s.validate.Struct(deck); err != nil {
+		return "", errors.ErrInvalidDeck
 	}
 
 	id, err := s.repo.AddDeck(ctx, deck)
@@ -57,24 +59,62 @@ func (s *DeckService) GetOneDeck(ctx context.Context, id string) (models.DeckRes
 
 	return models.DeckResponse{
 		ID:           id,
+		Title:        deck.Title,
 		OwnerID:      deck.OwnerID,
 		SharedEmails: deck.SharedEmails,
 		Cards:        cards,
 	}, nil
 }
 
-func (s *DeckService) UpdateDeck(ctx context.Context, deckID string, update models.UpdateDeck) error {
-	switch update.Operation {
-	case "add":
-		for _, id := range update.CardID {
-			return s.repo.AddCardToDeck(ctx, deckID, id)
-		}
-	case "remove":
-		for _, id := range update.CardID {
-			return s.repo.RemoveCardFromDeck(ctx, deckID, id)
+func (s *DeckService) UpdateDeck(
+	ctx context.Context,
+	deckID string,
+	update models.UpdateDeck,
+) (models.DeckResponse, error) {
+	var err error
+
+	if err := s.validate.Struct(update); err != nil {
+		return models.DeckResponse{}, errors.ErrInvalidDeck
+	}
+	if len(update.Cards) > 0 && update.Operation != "" {
+		for _, id := range update.Cards {
+			switch update.Operation {
+			case "add":
+				err = s.repo.AddCardToDeck(ctx, deckID, id)
+			case "remove":
+				err = s.repo.RemoveCardFromDeck(ctx, deckID, id)
+			default:
+				return models.DeckResponse{}, errors.ErrInvalidDeck
+			}
+
+			if err != nil {
+				return models.DeckResponse{}, err
+			}
 		}
 	}
-	return errors.ErrInvalidUser
+
+	updates, err := utils.StructToUpdate(update)
+	if err != nil {
+		return models.DeckResponse{}, err
+	}
+
+	filterd := make([]firestore.Update, 0, len(updates))
+	for _, u := range updates {
+		switch u.Path {
+		case "cards", "operation":
+			continue
+		default:
+			filterd = append(filterd, u)
+		}
+	}
+
+	if len(filterd) > 0 {
+		if err := s.repo.UpdateDeck(ctx, filterd, deckID); err != nil {
+			return models.DeckResponse{}, err
+		}
+	}
+
+	return s.GetOneDeck(ctx, deckID)
 }
 
 func getCardStructFromData(m map[string]any, errorOnFail error) (models.Card, error) {
